Use a single implementation of QueryResultHandler instead of 3
