+++ src/main/java/io/reactiverse/pgclient/impl/PgClientBase.java
@@ -30,7 +30,7 @@ public abstract class PgClientBase<C extends PgClient> implements PgClient {
 
   @Override
   public C query(String sql, Handler<AsyncResult<PgResult<PgRowSet>>> handler) {
-    schedule(new SimpleQueryCommand<>(sql, PgRowSetImpl.COLLECTOR, new SimpleQueryResultHandler<>(handler)));
+    schedule(new SimpleQueryCommand<>(sql, PgRowSetImpl.COLLECTOR, new PgResultBuilder<>(handler)));
     return (C) this;
   }
 
@@ -38,7 +38,7 @@ public abstract class PgClientBase<C extends PgClient> implements PgClient {
   public C preparedQuery(String sql, Tuple arguments, Handler<AsyncResult<PgResult<PgRowSet>>> handler) {
     schedule(new PrepareStatementCommand(sql, ar -> {
       if (ar.succeeded()) {
-        schedule(new ExtendedQueryCommand<>(ar.result(), arguments, PgRowSetImpl.COLLECTOR, new ExtendedQueryResultHandler<>(handler)));
+        schedule(new ExtendedQueryCommand<>(ar.result(), arguments, PgRowSetImpl.COLLECTOR, new PgResultBuilder<>(handler)));
       } else {
         handler.handle(Future.failedFuture(ar.cause()));
       }
@@ -59,7 +59,7 @@ public abstract class PgClientBase<C extends PgClient> implements PgClient {
           ar.result(),
           batch.iterator(),
           PgRowSetImpl.COLLECTOR,
-          new BatchQueryResultHandler(handler)));
+          new PgResultBuilder<>(handler)));
       } else {
         handler.handle(Future.failedFuture(ar.cause()));
       }
